\name{classifyDecisionTable-methods}
\docType{methods}
\alias{classifyDecisionTable-methods}
\alias{classifyDecisionTable,ValueReduct-method}
\title{Methods for Function \code{classifyDecisionTable}}
\description{
 Methods for function \code{classifyDecisionTable}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ValueReduct")}}{
This method uses the reduced rules in a Value Reduct object for classifying the rules in a Decision Table object. It returns a Decision Table object with the same original rule conditions but with the decision of the Value Reduct object which apply.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
