\name{computeCore-methods}
\docType{methods}
\alias{computeCore-methods}
\alias{computeCore,DiscernibilityMatrix-method}
\title{Methods for Function \code{computeCore}}
\description{
 Methods for function \code{computeCore}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DiscernibilityMatrix")}}{
This method computes the core of a Decision Table object; the core is computed using the discernibility matrix method. The core a set of conditions which are present in all the condition reducts of a decision table; in other words the core is the intersections of the conditions of all the condition reducts in a decision table.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
