\name{computeDiscernibilityMatrix-methods}
\docType{methods}
\alias{computeDiscernibilityMatrix-methods}
\alias{computeDiscernibilityMatrix,DecisionTable-method}
\title{Methods for Function \code{computeDiscernibilityMatrix}}
\description{
 Methods for function \code{computeDiscernibilityMatrix}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method computes the discernibility matrix of a Decision Table object. A Discernibility Matrix object is a 3 dimension boolean array where each element represents if there is a difference in the same condition of each pair of rules in a Decision Table object. A discernibility matrix is useful for calculating the core of a decision table.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
