\name{computeValueReduct-methods}
\docType{methods}
\alias{computeValueReduct-methods}
\alias{computeValueReduct,ConditionReduct-method}
\title{Methods for Function \code{computeValueReduct}}
\description{
 Methods for function \code{computeValueReduct}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
This method takes a Condition Reduct object and reduces its rules at the condition level and returns a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
