\name{findAllReductsFromCore-methods}
\docType{methods}
\alias{findAllReductsFromCore-methods}
\alias{findAllReductsFromCore,DecisionTable-method}
\title{Methods for Function \code{findAllReductsFromCore}}
\description{
 Methods for function \code{findAllReductsFromCore}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method seeks all the condition reducts in a Decision Table object. For doing this, it uses the core as an starting point and adds condition combinations until all the conditions in the Decision Table object has been added.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
