\name{findFirstConditionReduct-methods}
\docType{methods}
\alias{findFirstConditionReduct-methods}
\alias{findFirstConditionReduct,DecisionTable-method}
\title{Methods for Function \code{findFirstConditionReduct}}
\description{
 Methods for function \code{findFirstConditionReduct}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method returns the smallest first found condition reduct in a Decision Table object. For doing this, it uses the core as a starting point and adds condition combinations until it obtains a condition reduct.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
