\name{getColumnIds-methods}
\docType{methods}
\alias{getColumnIds-methods}
\alias{getColumnIds,ConditionReduct-method}
\title{Methods for Function \code{getColumnIds}}
\description{
 Methods for function \code{getColumnIds}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
This method returns the column ids of the conditions of a Decision Table object which are thought of being part of a condition reduct. To be sure if it really is a condition reduct the isConditionReduct method could be used.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
