\name{getCondition-methods}
\docType{methods}
\alias{getCondition-methods}
\alias{getCondition,DecisionTable-method}
\title{Methods for Function \code{getCondition}}
\description{
 Methods for function \code{getCondition}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method returns the conditions of all rules in a Decision Table object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}