\name{getConditionReduct-methods}
\docType{methods}
\alias{getConditionReduct-methods}
\alias{getConditionReduct,ConditionReduct-method}
\title{Methods for Function \code{getConditionReduct}}
\description{
 Methods for function \code{getConditionReduct}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
This method returns a numeric matrix as a representation of a Condition Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}
