\name{getConditionReductDecisionTable-methods}
\docType{methods}
\alias{getConditionReductDecisionTable-methods}
\alias{getConditionReductDecisionTable,ConditionReduct-method}
\title{Methods for Function \code{getConditionReductDecisionTable}}
\description{
 Methods for function \code{getConditionReductDecisionTable}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
This method returns the Decision Table object of a Condition Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}