\name{getDecision-methods}
\docType{methods}
\alias{getDecision-methods}
\alias{getDecision,DecisionTable-method}
\title{Methods for Function \code{getDecision}}
\description{
 Methods for function \code{getDecision}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method returns a numeric vector as a representation of the decision of the rules in a Decision Table object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}