\name{getDecisionTable-methods}
\docType{methods}
\alias{getDecisionTable-methods}
\alias{getDecisionTable,DecisionTable-method}
\title{Methods for Function \code{getDecisionTable}}
\description{
 Methods for function \code{getDecisionTable}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method returns a numeric matrix as a representation of a Decision Table object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}