\name{getDiscernibilityMatrix-methods}
\docType{methods}
\alias{getDiscernibilityMatrix-methods}
\alias{getDiscernibilityMatrix,DiscernibilityMatrix-method}
\title{Methods for Function \code{getDiscernibilityMatrix}}
\description{
 Methods for function \code{getDiscernibilityMatrix}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DiscernibilityMatrix")}}{
This method returns a 3 dimension boolean array representing a Discernibility Matrix object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}