\name{getRule-methods}
\docType{methods}
\alias{getRule-methods}
\alias{getRule,DecisionTable-method}
\title{Methods for Function \code{getRule}}
\description{
 Methods for function \code{getRule}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method returns a set of rules from a Decision Table object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}