\name{getValueReduct-methods}
\docType{methods}
\alias{getValueReduct-methods}
\alias{getValueReduct,ValueReduct-method}
\title{Methods for Function \code{getValueReduct}}
\description{
 Methods for function \code{getValueReduct}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ValueReduct")}}{
This method returns a numeric matrix representing a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}