\name{getValueReductConditionReduct-methods}
\docType{methods}
\alias{getValueReductConditionReduct-methods}
\alias{getValueReductConditionReduct,ValueReduct-method}
\title{Methods for Function \code{getValueReductConditionReduct}}
\description{
 Methods for function \code{getValueReductConditionReduct}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ValueReduct")}}{
This method returns the Condition Reduct object of a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}