\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,ConditionReduct-method}
\alias{initialize,DecisionTable-method}
\alias{initialize,DiscernibilityMatrix-method}
\alias{initialize,ValueReduct-method}
\title{Methods for Function \code{initialize}}
\description{
 Methods for function \code{initialize}
}
\section{Methods}{
\describe{

\item{\code{signature(.Object = "ConditionReduct")}}{
Constructor for a Condition Reduct object.
}

\item{\code{signature(.Object = "DecisionTable")}}{
Constructor for a Decision Table object.
}

\item{\code{signature(.Object = "DiscernibilityMatrix")}}{
Constructor for a Discernibility Matrix object.
}

\item{\code{signature(.Object = "ValueReduct")}}{
Constructor for a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}