\name{isConditionReduct-methods}
\docType{methods}
\alias{isConditionReduct-methods}
\alias{isConditionReduct,ConditionReduct-method}
\title{Methods for Function \code{isConditionReduct}}
\description{
 Methods for function \code{isConditionReduct}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
This method returns a boolean indicating if a Condition Reduct object actually is a condition reduct of its Decision Table object. 
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}