\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,ConditionReduct-method}
\alias{print,DecisionTable-method}
\alias{print,DiscernibilityMatrix-method}
\alias{print,ValueReduct-method}
\title{Methods for Function \code{print}}
\description{
 Methods for function \code{print}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ConditionReduct")}}{
It prints a Condition Reduct object.
}

\item{\code{signature(x = "DecisionTable")}}{
It prints a Decision Table object.
}

\item{\code{signature(x = "DiscernibilityMatrix")}}{
It prints a user friendly Discernibility Matrix object.
}

\item{\code{signature(x = "ValueReduct")}}{
It prints a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}