\name{removeDuplicatedRulesCR-methods}
\docType{methods}
\alias{removeDuplicatedRulesCR-methods}
\alias{removeDuplicatedRulesCR,ConditionReduct-method}
\title{Methods for Function \code{removeDuplicatedRulesCR}}
\description{
 Methods for function \code{removeDuplicatedRulesCR}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
This method removes the duplicated rules of a Condition Reduct object. For accomplishing this, the method removes rules from the Decision Table of the Condition Reduct object which are duplicated in the column ids that makes the condition reduct.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}