\name{removeDuplicatedRulesDT-methods}
\docType{methods}
\alias{removeDuplicatedRulesDT-methods}
\alias{removeDuplicatedRulesDT,DecisionTable-method}
\title{Methods for Function \code{removeDuplicatedRulesDT}}
\description{
 Methods for function \code{removeDuplicatedRulesDT}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method removes the duplicated rules in a Decision Table object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}