\name{removeDuplicatedRulesVR-methods}
\docType{methods}
\alias{removeDuplicatedRulesVR-methods}
\alias{removeDuplicatedRulesVR,ValueReduct-method}
\title{Methods for Function \code{removeDuplicatedRulesVR}}
\description{
 Methods for function \code{removeDuplicatedRulesVR}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ValueReduct")}}{
This method removes the duplicated rules in a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}