\name{rs-package}
\alias{rs-package}
\alias{rs}
\docType{package}
\title{
Simplification of Decision Tables using Rough Sets
}
\description{
Rough sets theory can be applied to reduce knowledge from decision tables. This package includes a few S4 classes for doing so.
}
\details{
\tabular{ll}{
Package: \tab rs\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-03-13\cr
License: \tab GPL (>= 2)\cr
Depends: \tab methods\cr
}
}
\author{
Alber Sanchez <alber.sanchez@uni-muenster.de>
}
\references{
Pawlak, Zdzislaw 1991 \emph{Rough Sets: Theoretical Aspects of Reasoning About Data} Dordrecht: Kluwer Academic Publishing.
}
\keyword{package}
\keyword{logic}
\keyword{rough}
\keyword{set}
