\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ConditionReduct-method}
\alias{show,DecisionTable-method}
\alias{show,DiscernibilityMatrix-method}
\alias{show,ValueReduct-method}
\title{Methods for Function \code{show}}
\description{
 Methods for function \code{show}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ConditionReduct")}}{
It shows the first ten rules and conditions of a Condition Reduct object.
}

\item{\code{signature(object = "DecisionTable")}}{
It shows the first ten rules and conditions of a Decision Table object.
}

\item{\code{signature(object = "DiscernibilityMatrix")}}{
It shows the first ten rows and columns of a Discernibility Matrix object.
}

\item{\code{signature(object = "ValueReduct")}}{
It shows the first ten rules and conditions of a Value Reduct object.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}