\name{simplifyDecisionTable-methods}
\docType{methods}
\alias{simplifyDecisionTable-methods}
\alias{simplifyDecisionTable,DecisionTable-method}
\title{Methods for Function \code{simplifyDecisionTable}}
\description{
 Methods for function \code{simplifyDecisionTable}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DecisionTable")}}{
This method returns a Value Reduct object. It is a shortcut for finding the first smallest condition reduct and after the value reduct.
}
}}
\keyword{methods}
\keyword{logic}
\keyword{rough}
\keyword{set}